Minor QDPlot module refactoring
