+++ src/qudi/gui/qdplot/main_window.py
